Moves Dto tests and Manifest from core
